--- client/src/Home.js
@@ -1,12 +1,0 @@
-import MainContainer from './containers/MainContainer';
-
-function Home() {
-  return (
-    <>
-        <MainContainer/>
-    </>
-    
-  );
-}
-export default Home;
-
