+++ client/src/NotFound.js
@@ -1,0 +1,9 @@
+const NotFound = () => {
+    return ( 
+        <>
+        <h2>Not Found</h2>
+        </>
+    );
+}
+ 
+export default NotFound;
